+++ app/uk/gov/hmrc/eventhub/model/Subscriber.scala
@@ -87,15 +87,3 @@ object Subscriber {
 
   implicit val fmt: OFormat[Subscriber] = Json.format[Subscriber]
 }
-
-case class Topic(name: String, subscribers: List[Subscriber])
-
-object Topic {
-  implicit val configLoader: ConfigLoader[Map[String, Topic]] = (rootConfig: Config, path: String) =>
-    ConfigSource
-      .fromConfig(rootConfig.getConfig(path))
-      .load[Map[String, Topic]] match {
-      case Left(value) => throw new IllegalArgumentException(s"could not load topic config: ${value.toList.mkString(" | ")}")
-      case Right(value) => value
-    }
-}
